fix: 无配置文件时报错

+++ src/MaaWpfGui/Configuration/ConfigFactory.cs
@@ -76,8 +76,12 @@ namespace MaaWpfGui.Configuration
 
                 if (parsed is null)
                 {
-                    _logger.Warning("Failed to load configuration file, copying configuration file to error file");
-                    File.Copy(_configurationFile, _configurationErrorFile, true);
+                    if (File.Exists(_configurationFile))
+                    {
+                        _logger.Warning("Failed to load configuration file, copying configuration file to error file");
+                        File.Copy(_configurationFile, _configurationErrorFile, true);
+                    }
+
                     if (File.Exists(_configurationBakFile))
                     {
                         _logger.Information("trying to use backup file");
