+++ src/MaaWpfGui/Helper/ConfigurationHelper.cs
@@ -183,20 +183,22 @@ namespace MaaWpfGui.Helper
                 return true;
             }
 
-            _logger.Warning("Failed to load configuration file, copying configuration file to error file");
-            File.Copy(_configurationFile, _configurationErrorFile, true);
+            if (File.Exists(_configurationFile))
+            {
+                _logger.Warning("Failed to load configuration file, copying configuration file to error file");
+                File.Copy(_configurationFile, _configurationErrorFile, true);
+            }
 
             if (File.Exists(_configurationBakFile))
             {
                 _logger.Information("trying to use backup file");
                 parsed = ParseJsonFile(_configurationBakFile);
-            }
-
-            if (parsed is not null)
-            {
-                _logger.Information("Backup file loaded successfully, copying backup file to configuration file");
-                File.Copy(_configurationBakFile, _configurationFile, true);
-                return true;
+                if (parsed is not null)
+                {
+                    _logger.Information("Backup file loaded successfully, copying backup file to configuration file");
+                    File.Copy(_configurationBakFile, _configurationFile, true);
+                    return true;
+                }
             }
 
             _logger.Warning("Failed to load configuration file and/or backup file, using default settings");
